ADAPT-89: Membership card links and logged out card
